add registration& delete &edit api

+++ App/components/Alert/AlertComp.tsx
@@ -2,11 +2,13 @@ import React from 'react';
 import {useNavigation} from '@react-navigation/native';
 import {Alert} from 'react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import {deleteAccount, Logout} from '../../api/Login/Login';
 export const AlertComp = (
   title,
   buttonTextLeft,
   buttonTextRight,
   navigation?,
+  handle?,
 ) => {
   // const navigation = useNavigation();
   // console.log(title);
@@ -15,8 +17,30 @@ export const AlertComp = (
       text: `${buttonTextLeft}`,
       onPress: () => {
         console.log('wwww');
-        navigation && navigation.navigate('Login');
-        navigation && AsyncStorage.clear();
+        if (handle == 'delete') {
+          navigation &&
+            deleteAccount()
+              .then(() => {
+                console.log('success')
+                navigation.navigate('Login');
+                navigation && AsyncStorage.clear();
+              })
+              .catch(err => {
+                console.log(err.response);
+                Alert.alert(err.message);
+              });
+        } else {
+          navigation &&
+            Logout()
+              .then(() => {
+                navigation.navigate('Login');
+                navigation && AsyncStorage.clear();
+              })
+              .catch(err => {
+                console.log(err.response);
+                Alert.alert(err.message);
+              });
+        }
       },
     },
     {text: `${buttonTextRight}`, onPress: () => console.log('OK Pressed')},
